HPA-17: tests for JpaSearchRepository
 - added tag and search to DatabaseTestSetup
 - added SearchEntityMatcher
 - changed searchCount in SearchEntity to a long

+++ headsplugin-rebuild/src/main/java/com/github/cc007/headsplugin/integration/database/entities/SearchEntity.java
@@ -54,7 +54,7 @@ public class SearchEntity {
 
     @Column(name = "searchCount")
     @Setter(AccessLevel.NONE)
-    private int searchCount;
+    private long searchCount;
 
     @Column(name = "lastUpdated")
     @Convert(converter = LocalDateTimeConverter.class)
